requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in
#
absl-py==1.0.0
    # via
    #   jax
    #   jaxlib
    #   tensorboard
aiohttp==3.8.1
    # via fsspec
aiosignal==1.2.0
    # via aiohttp
argon2-cffi==21.3.0
    # via notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asttokens==2.0.5
    # via stack-data
async-timeout==4.0.2
    # via aiohttp
attrs==21.4.0
    # via
    #   aiohttp
    #   jsonschema
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via nbconvert
bleach==5.0.0
    # via nbconvert
cachetools==5.0.0
    # via google-auth
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via argon2-cffi-bindings
charset-normalizer==2.0.12
    # via
    #   aiohttp
    #   requests
colorama==0.4.4
    # via unew
cycler==0.11.0
    # via matplotlib
debugpy==1.6.0
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
docopt==0.6.2
    # via unew
entrypoints==0.4
    # via
    #   jupyter-client
    #   nbconvert
executing==0.8.3
    # via stack-data
fastjsonschema==2.15.3
    # via nbformat
flatbuffers==2.0
    # via jaxlib
fonttools==4.32.0
    # via matplotlib
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2022.3.0
    # via pytorch-lightning
future==0.18.2
    # via unew
git-python==1.0.3
    # via -r requirements/requirements.in
gitdb==4.0.9
    # via gitpython
gitpython==3.1.27
    # via git-python
google-auth==2.6.5
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
grpcio==1.44.0
    # via tensorboard
idna==3.3
    # via
    #   requests
    #   yarl
importlib-metadata==4.11.3
    # via markdown
importlib-resources==5.7.0
    # via jsonschema
ipykernel==6.13.0
    # via
    #   -r requirements/requirements.in
    #   ipywidgets
    #   notebook
ipython==8.2.0
    # via
    #   ipykernel
    #   ipywidgets
ipython-genutils==0.2.0
    # via
    #   ipywidgets
    #   notebook
ipywidgets==7.7.0
    # via -r requirements/requirements.in
jax==0.3.7
    # via numpyro
jaxlib==0.3.7
    # via numpyro
jedi==0.18.1
    # via ipython
jinja2==3.1.1
    # via
    #   nbconvert
    #   notebook
jsonschema==4.4.0
    # via nbformat
jupyter-client==7.2.2
    # via
    #   ipykernel
    #   nbclient
    #   notebook
jupyter-core==4.9.2
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-widgets==1.1.0
    # via ipywidgets
kiwisolver==1.4.2
    # via matplotlib
lmdb==1.3.0
    # via -r requirements/requirements.in
markdown==3.3.6
    # via tensorboard
markupsafe==2.1.1
    # via
    #   jinja2
    #   nbconvert
matplotlib==3.5.1
    # via
    #   -r requirements/requirements.in
    #   seaborn
    #   unew
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
multipledispatch==0.6.0
    # via numpyro
nbclient==0.6.0
    # via nbconvert
nbconvert==6.5.0
    # via notebook
nbformat==5.3.0
    # via
    #   ipywidgets
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.5
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   notebook
notebook==6.4.10
    # via widgetsnbextension
numpy==1.22.3
    # via
    #   -r requirements/requirements.in
    #   jax
    #   jaxlib
    #   matplotlib
    #   numpyro
    #   opt-einsum
    #   pandas
    #   pyro-ppl
    #   pytorch-lightning
    #   scipy
    #   seaborn
    #   tensorboard
    #   torchmetrics
    #   unew
numpyro==0.9.2
    # via -r requirements/requirements.in
oauthlib==3.2.0
    # via requests-oauthlib
opt-einsum==3.3.0
    # via
    #   jax
    #   pyro-ppl
packaging==21.3
    # via
    #   ipykernel
    #   matplotlib
    #   nbconvert
    #   pytorch-lightning
    #   torchmetrics
pandas==1.4.2
    # via
    #   seaborn
    #   torch-tb-profiler
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.1.0
    # via matplotlib
prometheus-client==0.14.1
    # via notebook
prompt-toolkit==3.0.29
    # via ipython
protobuf==3.20.0
    # via tensorboard
psutil==5.9.0
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pydeprecate==0.3.2
    # via
    #   pytorch-lightning
    #   torchmetrics
pygments==2.11.2
    # via
    #   ipython
    #   nbconvert
pyparsing==3.0.8
    # via
    #   matplotlib
    #   packaging
pyro-api==0.1.2
    # via pyro-ppl
pyro-ppl==1.8.1
    # via -r requirements/requirements.in
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
pytorch-lightning==1.6.1
    # via -r requirements/requirements.in
pytz==2022.1
    # via pandas
pyyaml==6.0
    # via
    #   pytorch-lightning
    #   unew
pyzmq==22.3.0
    # via
    #   jupyter-client
    #   notebook
requests==2.27.1
    # via
    #   fsspec
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.8
    # via google-auth
scipy==1.8.0
    # via
    #   -r requirements/requirements.in
    #   jax
    #   jaxlib
    #   seaborn
    #   unew
seaborn==0.11.2
    # via -r requirements/requirements.in
send2trash==1.8.0
    # via notebook
six==1.16.0
    # via
    #   absl-py
    #   asttokens
    #   bleach
    #   google-auth
    #   grpcio
    #   multipledispatch
    #   python-dateutil
smmap==5.0.0
    # via gitdb
soupsieve==2.3.2.post1
    # via beautifulsoup4
stack-data==0.2.0
    # via ipython
tensorboard==2.8.0
    # via
    #   pytorch-lightning
    #   torch-tb-profiler
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
terminado==0.13.3
    # via notebook
tinycss2==1.1.1
    # via nbconvert
torch==1.11.0
    # via
    #   -r requirements/requirements.in
    #   pyro-ppl
    #   pytorch-lightning
    #   torchmetrics
torch-tb-profiler==0.4.0
    # via -r requirements/requirements.in
torchmetrics==0.8.0
    # via pytorch-lightning
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
tqdm==4.64.0
    # via
    #   numpyro
    #   pyro-ppl
    #   pytorch-lightning
    #   unew
traitlets==5.1.1
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
typing-extensions==4.1.1
    # via
    #   jax
    #   pytorch-lightning
    #   torch
unew @ git+https://bitbucket.org/betocollaboration/unew.git
    # via -r requirements/requirements.in
urllib3==1.26.9
    # via requests
voluptuous==0.13.1
    # via unew
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
werkzeug==2.1.1
    # via tensorboard
wheel==0.37.1
    # via tensorboard
widgetsnbextension==3.6.0
    # via ipywidgets
yarl==1.7.2
    # via aiohttp
zipp==3.8.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
